make readv and writev return ssize_t's
SCCS-vsn: lib/libc/sys/read.2 8.3
SCCS-vsn: lib/libc/sys/write.2 8.3
SCCS-vsn: sys/sys/uio.h 8.4

+++ usr/src/sys/sys/uio.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)uio.h	8.3 (Berkeley) %G%
+ *	@(#)uio.h	8.4 (Berkeley) %G%
  */
 
 #ifndef _SYS_UIO_H_
@@ -52,8 +52,8 @@ struct uio {
 #include <sys/cdefs.h>
 
 __BEGIN_DECLS
-int	readv __P((int, const struct iovec *, int));
-int	writev __P((int, const struct iovec *, int));
+ssize_t	readv __P((int, const struct iovec *, int));
+ssize_t	writev __P((int, const struct iovec *, int));
 __END_DECLS
 #endif /* !KERNEL */
 #endif /* !_SYS_UIO_H_ */
